Add new base page for migrated listing pages
