+++ tests/UI/pages/BO/customerService/orderMessages/index.js
@@ -1,50 +1,16 @@
 require('module-alias/register');
-const BOBasePage = require('@pages/BO/BObasePage');
+const MigratedListingBasePage = require('@pages/BO/migratedListingBasePage.js');
 
-class OrderMessages extends BOBasePage {
+const tableName = 'order_message';
+
+class OrderMessages extends MigratedListingBasePage {
   constructor() {
-    super();
+    super(tableName);
 
     this.pageTitle = 'Order Messages •';
 
     // Selectors header
     this.newOrderMessageLink = '#page-header-desc-configuration-add';
-
-    // Selectors grid panel
-    this.gridPanel = '#order_message_grid_panel';
-    this.gridTable = '#order_message_grid_table';
-    this.gridHeaderTitle = `${this.gridPanel} h3.card-header-title`;
-    // Filters
-    this.filterColumn = filterBy => `${this.gridTable} #order_message_${filterBy}`;
-    this.filterSearchButton = `${this.gridTable} .grid-search-button`;
-    this.filterResetButton = `${this.gridTable} .grid-reset-button`;
-    // Table rows and columns
-    this.tableBody = `${this.gridTable} tbody`;
-    this.tableRow = row => `${this.tableBody} tr:nth-child(${row})`;
-    this.tableEmptyRow = `${this.tableBody} tr.empty_row`;
-    this.tableColumn = (row, column) => `${this.tableRow(row)} td.column-${column}`;
-    // Actions buttons in Row
-    this.actionsColumn = row => `${this.tableRow(row)} td.column-actions`;
-    this.editRowLink = row => `${this.actionsColumn(row)} a.grid-edit-row-link`;
-    this.dropdownToggleButton = row => `${this.actionsColumn(row)} a.dropdown-toggle`;
-    this.dropdownToggleMenu = row => `${this.actionsColumn(row)} div.dropdown-menu`;
-    this.deleteRowLink = row => `${this.dropdownToggleMenu(row)} a.grid-delete-row-link`;
-    // Bulk Actions
-    this.selectAllRowsLabel = `${this.gridPanel} tr.column-filters .grid_bulk_action_select_all`;
-    this.bulkActionsToggleButton = `${this.gridPanel} button.js-bulk-actions-btn`;
-    this.bulkActionsDeleteButton = '#order_message_grid_bulk_action_delete_selection';
-    // Delete modal
-    this.confirmDeleteModal = '#order_message-grid-confirm-modal';
-    this.confirmDeleteButton = `${this.confirmDeleteModal} button.btn-confirm-submit`;
-    // Pagination selectors
-    this.paginationLimitSelect = '#paginator_select_page_limit';
-    this.paginationLabel = `${this.gridPanel} .col-form-label`;
-    this.paginationNextLink = `${this.gridPanel} #pagination_next_url`;
-    this.paginationPreviousLink = `${this.gridPanel} [aria-label='Previous']`;
-    // Sort Selectors
-    this.tableHead = `${this.gridPanel} thead`;
-    this.sortColumnDiv = column => `${this.tableHead} div.ps-sortable-column[data-sort-col-name='${column}']`;
-    this.sortColumnSpanButton = column => `${this.sortColumnDiv(column)} span.ps-sort`;
   }
 
   /* Header Methods */
@@ -55,216 +21,6 @@ class OrderMessages extends BOBasePage {
    */
   async goToAddNewOrderMessagePage(page) {
     await this.clickAndWaitForNavigation(page, this.newOrderMessageLink);
-  }
-
-
-  /* Reset Methods */
-  /**
-   * Reset filters in table
-   * @param page
-   * @return {Promise<void>}
-   */
-  async resetFilter(page) {
-    if (await this.elementVisible(page, this.filterResetButton, 2000)) {
-      await this.clickAndWaitForNavigation(page, this.filterResetButton);
-    }
-  }
-
-  /**
-   * get number of elements in grid
-   * @param page
-   * @returns {Promise<number>}
-   */
-  async getNumberOfElementInGrid(page) {
-    return this.getNumberFromText(page, this.gridHeaderTitle);
-  }
-
-  /**
-   * Reset Filter And get number of elements in list
-   * @param page
-   * @returns {Promise<number>}
-   */
-  async resetAndGetNumberOfLines(page) {
-    await this.resetFilter(page);
-    return this.getNumberOfElementInGrid(page);
-  }
-
-  /* filter Methods */
-  /**
-   * Filter Table
-   * @param page
-   * @param filterBy, which column
-   * @param value, value to put in filter
-   * @return {Promise<void>}
-   */
-  async filterTable(page, filterBy, value) {
-    await this.setValue(page, this.filterColumn(filterBy), value);
-    await this.clickAndWaitForNavigation(page, this.filterSearchButton);
-  }
-
-  /* Column Methods */
-  /**
-   * Edit order message
-   * @param page
-   * @param row
-   * @return {Promise<void>}
-   */
-  async gotoEditOrderMessage(page, row = 1) {
-    await this.clickAndWaitForNavigation(page, this.editRowLink(row));
-  }
-
-  /**
-   * Delete Row in table
-   * @param page
-   * @param row, row to delete
-   * @returns {Promise<string>}
-   */
-  async deleteOrderMessage(page, row = 1) {
-    await Promise.all([
-      page.click(this.dropdownToggleButton(row)),
-      this.waitForVisibleSelector(
-        page,
-        `${this.dropdownToggleButton(row)}[aria-expanded='true']`,
-      ),
-    ]);
-
-    // Click on delete and wait for modal
-    await Promise.all([
-      page.click(this.deleteRowLink(row)),
-      this.waitForVisibleSelector(page, `${this.confirmDeleteModal}.show`),
-    ]);
-
-    // Confirm delete in modal
-    await this.confirmDeleteOrderMessages(page);
-    return this.getTextContent(page, this.alertSuccessBlockParagraph);
-  }
-
-  /**
-   * Get text from a column
-   * @param page
-   * @param row, row in table
-   * @param column, which column
-   * @returns {Promise<string>}
-   */
-  async getTextColumnFromTable(page, row, column) {
-    return this.getTextContent(page, this.tableColumn(row, column));
-  }
-
-  /* Bulk Actions Methods */
-  /**
-   * Delete with bulk actions
-   * @param page
-   * @returns {Promise<string>}
-   */
-  async deleteWithBulkActions(page) {
-    // Click on Select All
-    await Promise.all([
-      page.$eval(this.selectAllRowsLabel, el => el.click()),
-      this.waitForVisibleSelector(page, `${this.bulkActionsToggleButton}:not([disabled])`),
-    ]);
-    // Click on Button Bulk actions
-    await Promise.all([
-      page.click(this.bulkActionsToggleButton),
-      this.waitForVisibleSelector(page, `${this.bulkActionsToggleButton}[aria-expanded='true']`),
-    ]);
-    // Click on delete and wait for modal
-    await Promise.all([
-      page.click(this.bulkActionsDeleteButton),
-      this.waitForVisibleSelector(page, `${this.confirmDeleteModal}.show`),
-    ]);
-    // Click on delete and wait for modal
-    await this.confirmDeleteOrderMessages(page);
-    return this.getTextContent(page, this.alertSuccessBlockParagraph);
-  }
-
-  /**
-   * Confirm delete in modal
-   * @param page
-   * @return {Promise<void>}
-   */
-  async confirmDeleteOrderMessages(page) {
-    await this.clickAndWaitForNavigation(page, this.confirmDeleteButton);
-  }
-
-  /* Pagination methods */
-  /**
-   * Get pagination label
-   * @param page
-   * @returns {Promise<string>}
-   */
-  getPaginationLabel(page) {
-    return this.getTextContent(page, this.paginationLabel);
-  }
-
-  /**
-   * Select pagination limit
-   * @param page
-   * @param number
-   * @returns {Promise<string>}
-   */
-  async selectPaginationLimit(page, number) {
-    await this.selectByVisibleText(page, this.paginationLimitSelect, number);
-    return this.getPaginationLabel(page);
-  }
-
-  /**
-   * Click on next
-   * @param page
-   * @returns {Promise<string>}
-   */
-  async paginationNext(page) {
-    await this.clickAndWaitForNavigation(page, this.paginationNextLink);
-    return this.getPaginationLabel(page);
-  }
-
-  /**
-   * Click on previous
-   * @param page
-   * @returns {Promise<string>}
-   */
-  async paginationPrevious(page) {
-    await this.clickAndWaitForNavigation(page, this.paginationPreviousLink);
-    return this.getPaginationLabel(page);
-  }
-
-  // Sort methods
-  /**
-   * Get content from all rows
-   * @param page
-   * @param column
-   * @return {Promise<[]>}
-   */
-  async getAllRowsColumnContent(page, column) {
-    const rowsNumber = await this.getNumberOfElementInGrid(page);
-    const allRowsContentTable = [];
-    for (let i = 1; i <= rowsNumber; i++) {
-      let rowContent = await this.getTextContent(page, this.tableColumn(i, column));
-      if (column === 'active') {
-        rowContent = await this.getToggleColumnValue(page, i).toString();
-      }
-      await allRowsContentTable.push(rowContent);
-    }
-    return allRowsContentTable;
-  }
-
-  /**
-   * Sort table
-   * @param page
-   * @param sortBy, column to sort with
-   * @param sortDirection, asc or desc
-   * @return {Promise<void>}
-   */
-  async sortTable(page, sortBy, sortDirection = 'asc') {
-    const sortColumnDiv = `${this.sortColumnDiv(sortBy)}[data-sort-direction='${sortDirection}']`;
-    const sortColumnSpanButton = this.sortColumnSpanButton(sortBy);
-
-    let i = 0;
-    while (await this.elementNotVisible(page, sortColumnDiv, 2000) && i < 2) {
-      await this.clickAndWaitForNavigation(page, sortColumnSpanButton);
-      i += 1;
-    }
-
-    await this.waitForVisibleSelector(page, sortColumnDiv, 20000);
   }
 }
